+++ userspace/lib/src/posix_signal.c
@@ -1,9 +1,11 @@
 #include <errno.h>
 #include <signal.h>
 #include <srvros/sys.h>
+#include <stdint.h>
 #include <unistd.h>
 
-static sighandler_t signal_handlers[64];
+static struct sigaction signal_actions[64];
+static sigset_t process_signal_mask;
 
 static int signal_valid(int signum) {
     return signum > 0 && signum < (int)(sizeof(sigset_t) * 8);
@@ -13,12 +15,11 @@ static int signal_supported(int signum) {
     return signum == SIGINT || signum == SIGTERM;
 }
 
-sighandler_t signal(int signum, sighandler_t handler) {
-    if (!signal_valid(signum) || !signal_supported(signum)) {
-        errno = EINVAL;
-        return SIG_ERR;
-    }
+static sigset_t signal_mask_for(int signum) {
+    return (sigset_t)1 << signum;
+}
 
+static int signal_apply_action(int signum, sighandler_t handler) {
     uint64_t action = SRV_SIGNAL_DEFAULT;
     if (handler == SIG_IGN) {
         action = SRV_SIGNAL_IGNORE;
@@ -27,11 +28,22 @@ sighandler_t signal(int signum, sighandler_t handler) {
     }
     if (srv_signal_config((uint64_t)signum, action) < 0) {
         errno = EINVAL;
+        return -1;
+    }
+    return 0;
+}
+
+sighandler_t signal(int signum, sighandler_t handler) {
+    struct sigaction action = {
+        .sa_handler = handler,
+        .sa_mask = 0,
+        .sa_flags = 0,
+    };
+    struct sigaction old_action;
+    if (sigaction(signum, &action, &old_action) < 0) {
         return SIG_ERR;
     }
-    sighandler_t previous = signal_handlers[signum];
-    signal_handlers[signum] = handler;
-    return previous;
+    return old_action.sa_handler;
 }
 
 int kill(pid_t pid, int sig) {
@@ -51,7 +63,7 @@ int raise(int sig) {
         errno = EINVAL;
         return -1;
     }
-    sighandler_t handler = signal_handlers[sig];
+    sighandler_t handler = signal_actions[sig].sa_handler;
     if (handler == SIG_IGN) {
         return 0;
     }
@@ -60,6 +72,105 @@ int raise(int sig) {
         return 0;
     }
     return kill(getpid(), sig);
+}
+
+int sigaction(int signum, const struct sigaction *act, struct sigaction *oldact) {
+    if (!signal_valid(signum) || !signal_supported(signum)) {
+        errno = EINVAL;
+        return -1;
+    }
+    if (act != 0 && (act->sa_flags & ~SA_RESTART) != 0) {
+        errno = EINVAL;
+        return -1;
+    }
+
+    if (oldact != 0) {
+        *oldact = signal_actions[signum];
+    }
+    if (act != 0) {
+        if (signal_apply_action(signum, act->sa_handler) < 0) {
+            return -1;
+        }
+        signal_actions[signum] = *act;
+    }
+    return 0;
+}
+
+int sigprocmask(int how, const sigset_t *set, sigset_t *oldset) {
+    if (oldset != 0) {
+        *oldset = process_signal_mask;
+    }
+    if (set == 0) {
+        return 0;
+    }
+
+    switch (how) {
+    case SIG_BLOCK:
+        process_signal_mask |= *set;
+        return 0;
+    case SIG_UNBLOCK:
+        process_signal_mask &= ~(*set);
+        return 0;
+    case SIG_SETMASK:
+        process_signal_mask = *set;
+        return 0;
+    default:
+        errno = EINVAL;
+        return -1;
+    }
+}
+
+int pthread_sigmask(int how, const sigset_t *set, sigset_t *oldset) {
+    int saved_errno = errno;
+    if (sigprocmask(how, set, oldset) < 0) {
+        int error = errno != 0 ? errno : EINVAL;
+        errno = saved_errno;
+        return error;
+    }
+    errno = saved_errno;
+    return 0;
+}
+
+int sigpending(sigset_t *set) {
+    if (set == 0) {
+        errno = EINVAL;
+        return -1;
+    }
+
+    uint64_t pending = 0;
+    if (srv_signal_pending(&pending) < 0) {
+        errno = EINVAL;
+        return -1;
+    }
+    *set = (sigset_t)pending & process_signal_mask;
+    return 0;
+}
+
+int sigwait(const sigset_t *set, int *sig) {
+    if (set == 0 || sig == 0) {
+        return EINVAL;
+    }
+
+    for (;;) {
+        uint64_t pending = 0;
+        if (srv_signal_pending(&pending) < 0) {
+            return EINVAL;
+        }
+        uint64_t wanted = pending & (uint64_t)(*set);
+        for (int signum = 1; signum < 64; signum++) {
+            if ((wanted & (uint64_t)signal_mask_for(signum)) != 0) {
+                uint64_t polled = 0;
+                do {
+                    if (srv_signal_poll(&polled) < 0) {
+                        return EINVAL;
+                    }
+                } while (polled != 0 && polled != (uint64_t)signum);
+                *sig = signum;
+                return 0;
+            }
+        }
+        srv_yield();
+    }
 }
 
 int sigemptyset(sigset_t *set) {